+++ R/check_gauss_kruger.R
@@ -1,3 +1,12 @@
+#' Calculate the 3-degree Gauss Kruger zone number and the central longitude of the input longitude range
+#'
+#' @param lonmax A numerical value, max longitude of the map extent
+#' @param lonmin A numerical value, min longitude of the map extent
+#'
+#' @returns A list with name of "zone_num" and "mid_lon", if the input range doesn't fall in any zone, the list will be NAs.
+#' @export
+#'
+#' @examples check_gauss_kruger_3_deg(11,12)
 check_gauss_kruger_3_deg = function(lonmax, lonmin) {
   zone_number_3 = \(lon) as.integer((lon + 1.5) / 3)
   N1 = zone_number_3(lonmax)
@@ -10,6 +19,16 @@ check_gauss_kruger_3_deg = function(lonmax, lonmin) {
   }
 }
 
+
+#' Calculate the 6-degree Gauss Kruger zone number and the central longitude of the input longitude range
+#'
+#' @param lonmax A numerical value, max longitude of the map extent
+#' @param lonmin A numerical value, min longitude of the map extent
+#'
+#' @returns A list with name of "zone_num" and "mid_lon", if the input range doesn't fall in any zone, the list will be NAs.
+#' @export
+#'
+#' @examples check_gauss_kruger_6_deg(11,12)
 check_gauss_kruger_6_deg = function(lonmax, lonmin) {
   zone_number_6 = \(lon) as.integer((lon + 6) / 6)
   N1 = zone_number_6(lonmax)