+++ news.md
@@ -1,3 +1,8 @@
+# projWiz 0.2.0
+- remove dependence on geosphere
+- remove conic check, instead use extent range check
+- smaller region auto azimuthal projection
+
 # projWiz 0.1.2
 
 - documentation and examples update